Refactor: move data loader and data manager from playback module to database module
